CNV-76092: fix switching Checkups tab on project change
- also fixes hidden MigrationPolicies title
